pyproject.toml
[tool.poetry]
name = "kod"
version = "1.4.5"
description = "Python/Node.js project"
authors = ["Oleg Kizyma <oleg1203@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.10"
prometheus_client = "^0.15.0"
fastapi = "^0.109.0"  # Обновлено
pydantic = "^2.0.0"   # Добавлено
uvicorn = "^0.21.0"
google-generativeai = "^0.8.3"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
pytest-watch = "^4.2.0"
flake8 = "^6.0.0"
mypy = "^1.3.0"
black = "^23.3.0"
autopep8 = "^2.0.0"
pylint = "^2.17.0"
autoflake = "^2.1.0"
rope = "^1.7.0"
pyright = "^1.1.0"
watchdog = "^3.0.0"
types-requests = "^2.31.0.20240311"
types-setuptools = "68.2.0.2"
types-six = "^1.16.21"
types-python-dateutil = "^2.8.19"
types-pytz = "^2023.3"
types-pyyaml = "^6.0.12"
mypy-extensions = "1.0.0"
typing-extensions = "4.8.0"

[tool.poetry.scripts]
start = "src.python_app.main:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.10"
strict = true

[tool.autopep8]
max_line_length = 100
aggressive = 3

[tool.pylint]
max-line-length = 100
disable = ["C0111", "C0103"]

[tool.pyright]
typeCheckingMode = "strict"

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=src"

[tool.black]
line-length = 100
target-version = ["py310"]
include = '\.pyi?$'

[tool.flake8]
max-line-length = 100
exclude = [".git", "__pycache__", "build", "dist"]
